campusやdivisionもcontextで取得できるように
