+++ web/src/app/(auth)/settings/language/page.tsx
@@ -128,7 +128,7 @@ export default function Page() {
         ...formData,
         guid: user.uid,
       };
-      const res = await client.users.me.$put({ json: body });
+      const res = await client.users.me.$patch({ json: body });
       if (!res.ok) {
         console.error(await res.text());
         throw new Error(`レスポンスステータス: ${res.status}`);
